Start work on passing --server arg
