+++ browser/src/index.tsx
@@ -266,7 +266,7 @@ const start = (args: string[]) => {
     configChange() // initialize values
     config.registerListener(configChange)
 
-    instance.start(parsedArgs._)
+    instance.start(parsedArgs)
 
     const mouse = new Mouse(editorElement, screen)
